requirements.txt
livekit-agents[openai,google,silero,deepgram]>=1.0.0
openai>=1.40.0
chromadb>=0.5.15
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
google-cloud-texttospeech>=2.17.0
google-cloud-speech>=2.27.0
python-dotenv>=1.0.1
pydantic>=2.7.4
tiktoken>=0.7.0
tenacity>=8.4.0
loguru>=0.7.2
pyyaml>=6.0.0
